User accounts and onboarding flow (#3)
* Switch to uint and stage refactor with user model

* Cleanup

* Check for bot user ID in new conversations

* Enable user onboarding flow

* User manager

* Test GetResponses

* Remove success logs

+++ test/mocks.go
@@ -7,7 +7,7 @@ import (
 	"time"
 )
 
-const TestUserId = 12345
+const TestUserId = uint(12345)
 
 type TelegramMock struct {
 	Calls struct {
